feat(charts): new chart ui for continuous questions
