+++ front_end/src/app/(main)/(tournaments)/tournament/components/index/index_timeline.tsx
@@ -3,7 +3,7 @@ import { useTranslations } from "next-intl";
 import { FC, useCallback, useState } from "react";
 import { DomainTuple } from "victory";
 
-import NewNumericChart from "@/components/charts/new_numeric_chart";
+import NumericChart from "@/components/charts/numeric_chart";
 import { METAC_COLORS } from "@/constants/colors";
 import { Area, Line } from "@/types/charts";
 import { TimelineChartZoomOption } from "@/types/charts";
@@ -33,7 +33,7 @@ const IndexTimeline: FC<Props> = ({ tournament, height = 170 }) => {
   };
 
   return (
-    <NewNumericChart
+    <NumericChart
       buildChartData={buildChartData}
       withZoomPicker={true}
       height={height}
