+++ front_end/src/app/(main)/aggregation-explorer/components/aggregation_tab.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import { saveAs } from "file-saver";
+import { isNil } from "lodash";
 import { useTranslations } from "next-intl";
 import { FC, useCallback, useState, memo, useMemo } from "react";
 import toast from "react-hot-toast";
 
-import NumericChart from "@/components/charts/numeric_chart";
 import NumericTimeline from "@/components/charts/numeric_timeline";
 import DetailsQuestionCardErrorBoundary from "@/components/detailed_question_card/detailed_question_card/error_boundary";
 import CursorDetailItem from "@/components/detailed_question_card/detailed_question_card/numeric_cursor_item";
@@ -85,14 +85,17 @@ const AggregationsTab: FC<Props> = ({
       return null;
     }
 
-    const index = activeAggregation.history.findIndex(
-      (f) => f.start_time === cursorTimestamp
-    );
+    const index = isNil(cursorTimestamp)
+      ? -1
+      : activeAggregation.history.findLastIndex(
+          (f) => f.start_time <= cursorTimestamp
+        );
 
     const forecast =
       index === -1
         ? activeAggregation.history.at(-1)
         : activeAggregation.history[index];
+
     if (!forecast) {
       return null;
     }
@@ -203,29 +206,19 @@ const AggregationsTab: FC<Props> = ({
         </div>
       )}
       <DetailsQuestionCardErrorBoundary>
-        {aggregationData.type === QuestionType.Binary ? (
-          <NumericTimeline
-            aggregation={activeAggregation}
-            aggregationIndex={aggregationIndex}
-            questionType={aggregationData.type}
-            actualCloseTime={actualCloseTime}
-            scaling={aggregationData.scaling}
-            resolution={resolution}
-            onCursorChange={handleCursorChange}
-            unit={unit}
-          />
-        ) : (
-          <NumericChart
-            aggregation={activeAggregation}
-            aggregationIndex={aggregationIndex}
-            questionType={aggregationData.type}
-            actualCloseTime={actualCloseTime}
-            scaling={aggregationData.scaling}
-            resolution={resolution}
-            onCursorChange={handleCursorChange}
-            unit={unit}
-          />
-        )}
+        <NumericTimeline
+          aggregation={activeAggregation}
+          aggregationIndex={aggregationIndex}
+          questionType={aggregationData.type}
+          actualCloseTime={actualCloseTime}
+          scaling={aggregationData.scaling}
+          resolution={resolution}
+          cursorTimestamp={cursorTimestamp}
+          onCursorChange={handleCursorChange}
+          unit={unit}
+          simplifiedCursor={aggregationData.type !== QuestionType.Binary}
+        />
+
         {!!cursorData && (
           <div className="my-3 flex flex-wrap items-center justify-center gap-x-4 gap-y-2 xs:gap-x-8 sm:mx-8 sm:grid sm:grid-cols-2 sm:gap-x-4 sm:gap-y-0">
             <CursorDetailItem
