+++ front_end/src/components/detailed_question_card/detailed_question_card/continuous_chart_card.tsx
@@ -3,7 +3,6 @@ import { isNil } from "lodash";
 import { useTranslations } from "next-intl";
 import React, { FC, ReactNode, useCallback, useMemo, useState } from "react";
 
-import NumericChart from "@/components/charts/numeric_chart";
 import NumericTimeline from "@/components/charts/numeric_timeline";
 import CPRevealTime from "@/components/cp_reveal_time";
 import { useAuth } from "@/contexts/auth_context";
@@ -176,61 +175,33 @@ const DetailedContinuousChartCard: FC<Props> = ({
       )}
     >
       <div className="relative">
-        {question.type === QuestionType.Binary ? (
-          <NumericTimeline
-            aggregation={question.aggregations.recency_weighted}
-            myForecasts={question.my_forecasts}
-            resolution={question.resolution}
-            resolveTime={question.actual_resolve_time}
-            cursorTimestamp={cursorTimestamp}
-            onCursorChange={handleCursorChange}
-            onChartReady={handleChartReady}
-            questionType={question.type}
-            actualCloseTime={getPostDrivenTime(question.actual_close_time)}
-            scaling={question.scaling}
-            defaultZoom={
-              user
-                ? TimelineChartZoomOption.All
-                : TimelineChartZoomOption.TwoMonths
-            }
-            withZoomPicker
-            hideCP={hideCP || !!forecastAvailability?.cpRevealsOn}
-            isEmptyDomain={
-              !!forecastAvailability?.isEmpty ||
-              !!forecastAvailability?.cpRevealsOn
-            }
-            openTime={getPostDrivenTime(question.open_time)}
-            unit={question.unit}
-          />
-        ) : (
-          <NumericChart
-            aggregation={question.aggregations.recency_weighted}
-            myForecasts={question.my_forecasts}
-            resolution={question.resolution}
-            resolveTime={question.actual_resolve_time}
-            onCursorChange={handleCursorChange}
-            yLabel={t("communityPredictionLabel")}
-            onChartReady={handleChartReady}
-            questionType={question.type}
-            actualCloseTime={getPostDrivenTime(question.actual_close_time)}
-            scaling={question.scaling}
-            defaultZoom={
-              user
-                ? TimelineChartZoomOption.All
-                : TimelineChartZoomOption.TwoMonths
-            }
-            withZoomPicker
-            hideCP={hideCP || !!forecastAvailability?.cpRevealsOn}
-            withUserForecastTimestamps={!!forecastAvailability?.cpRevealsOn}
-            isEmptyDomain={
-              !!forecastAvailability?.isEmpty ||
-              !!forecastAvailability?.cpRevealsOn
-            }
-            openTime={getPostDrivenTime(question.open_time)}
-            unit={question.unit}
-            inboundOutcomeCount={question.inbound_outcome_count}
-          />
-        )}
+        <NumericTimeline
+          aggregation={question.aggregations.recency_weighted}
+          myForecasts={question.my_forecasts}
+          resolution={question.resolution}
+          resolveTime={question.actual_resolve_time}
+          cursorTimestamp={cursorTimestamp}
+          onCursorChange={handleCursorChange}
+          onChartReady={handleChartReady}
+          questionType={question.type}
+          actualCloseTime={getPostDrivenTime(question.actual_close_time)}
+          scaling={question.scaling}
+          defaultZoom={
+            user
+              ? TimelineChartZoomOption.All
+              : TimelineChartZoomOption.TwoMonths
+          }
+          withZoomPicker
+          hideCP={hideCP || !!forecastAvailability?.cpRevealsOn}
+          isEmptyDomain={
+            !!forecastAvailability?.isEmpty ||
+            !!forecastAvailability?.cpRevealsOn
+          }
+          openTime={getPostDrivenTime(question.open_time)}
+          unit={question.unit}
+          inboundOutcomeCount={question.inbound_outcome_count}
+          simplifiedCursor={question.type !== QuestionType.Binary}
+        />
       </div>
       <div
         className={cn(
