+++ front_end/src/components/forecast_card.tsx
@@ -1,11 +1,11 @@
 "use client";
+import { isNil } from "lodash";
 import Link from "next/link";
 import { FC, memo, useEffect, useMemo, useRef, useState } from "react";
 
 import { EmbedTheme } from "@/app/(embed)/questions/constants/embed_theme";
 import MultipleChoiceGroupChart from "@/app/(main)/questions/[id]/components/multiple_choice_group_chart";
 import FanChart from "@/components/charts/fan_chart";
-import NumericChart from "@/components/charts/numeric_chart";
 import NumericTimeline from "@/components/charts/numeric_timeline";
 import ConditionalTile from "@/components/conditional_tile";
 import DetailedMultipleChoiceChartCard from "@/components/detailed_question_card/detailed_question_card/multiple_choice_chart_card";
@@ -120,6 +120,7 @@ const ForecastCard: FC<Props> = ({
                 actualCloseTime={getPostDrivenTime(question.actual_close_time)}
                 scaling={question.scaling}
                 onCursorChange={nonInteractive ? undefined : setCursorTimestamp}
+                cursorTimestamp={cursorTimestamp}
                 nonInteractive={nonInteractive}
                 extraTheme={embedTheme?.chart}
                 defaultZoom={defaultChartZoom}
@@ -131,6 +132,8 @@ const ForecastCard: FC<Props> = ({
                 openTime={getPostDrivenTime(question.open_time)}
                 unit={question.unit}
                 tickFontSize={9}
+                isEmbedded={!isNil(embedTheme)}
+                simplifiedCursor={!isNil(embedTheme)}
               />
               <ForecastAvailabilityChartOverflow
                 forecastAvailability={forecastAvailability}
@@ -139,8 +142,8 @@ const ForecastCard: FC<Props> = ({
             </div>
           );
         case QuestionType.Numeric:
-        case QuestionType.Discrete:
         case QuestionType.Date:
+        case QuestionType.Discrete:
           if (embedChartType === EmbedChartType.Current) {
             return (
               <div className="flex w-full flex-col">
@@ -160,7 +163,7 @@ const ForecastCard: FC<Props> = ({
           } else {
             return (
               <div className="relative flex w-full flex-col">
-                <NumericChart
+                <NumericTimeline
                   aggregation={question.aggregations.recency_weighted}
                   myForecasts={question.my_forecasts}
                   resolution={question.resolution}
@@ -174,18 +177,20 @@ const ForecastCard: FC<Props> = ({
                   onCursorChange={
                     nonInteractive ? undefined : setCursorTimestamp
                   }
+                  cursorTimestamp={cursorTimestamp}
+                  nonInteractive={nonInteractive}
                   extraTheme={embedTheme?.chart}
                   defaultZoom={defaultChartZoom}
                   withZoomPicker={withZoomPicker}
-                  withUserForecastTimestamps={
-                    !!forecastAvailability.cpRevealsOn
-                  }
                   isEmptyDomain={
                     forecastAvailability.isEmpty ||
                     !!forecastAvailability.cpRevealsOn
                   }
                   openTime={getPostDrivenTime(question.open_time)}
                   unit={question.unit}
+                  tickFontSize={9}
+                  simplifiedCursor={true}
+                  isEmbedded={!isNil(embedTheme)}
                   inboundOutcomeCount={question.inbound_outcome_count}
                 />
                 <ForecastAvailabilityChartOverflow
